pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "berryql"
version = "0.1.0"
description = "A powerful GraphQL query optimization library for Strawberry GraphQL and SQLAlchemy"
readme = "README.md"
license = "MIT"
authors = [
    {name = "BerryQL Contributors"},
]
keywords = ["graphql", "strawberry", "sqlalchemy", "query-optimization", "orm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Framework :: AsyncIO",
]
requires-python = ">=3.8"
dependencies = [
    "strawberry-graphql>=0.200.0",
    "sqlalchemy>=2.0.0",
    "graphql-core>=3.2.0",
    "inflection>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "flake8>=4.0.0",
    "pre-commit>=2.15.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "aiosqlite>=0.17.0",  # For testing with SQLite
]

[project.urls]
Homepage = "https://github.com/berryql/berryql"
Repository = "https://github.com/berryql/berryql"
Documentation = "https://berryql.readthedocs.io"
"Bug Tracker" = "https://github.com/berryql/berryql/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["berryql*"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
